Fix CORS error in bitcoin market data call
The API calls to CoinMarketCap are failing due to a
CORS error. CoinMarketCap does not support API calls
directly from a client's browser. This prevents the
current USD value of all BTC sold from displaying.

This PR replaces CoinMarketCap with CoinGecko - a reliable
alternative - which does allow client browsers to interact
with it. (CoinGecko is gaining popularity and also used
by default in BTCPayServer now)